Few improvements to the yaxis legend plugin.
It allows series to configure if it should be visible in the yaxis legend using its `showRects` option. Provides `showRectsX` and `showRectsY` to offset the legend. Stores `yaxisGroup` and `yaxisRect` into the chart and series, respectively, so that outside code can modify them if needed. Redraw on the `redraw` event, so that dynamically loaded graphs work. And re-bind event handlers so that for series which are loaded after initialization event handlers are bound as well.